requirements.txt
absl-py==2.1.0
ai21==2.9.1
ai21-tokenizer==0.11.2
aiohttp==3.9.0
aiosignal==1.3.1
anndata==0.9.2
annotated-types==0.7.0
anthropic==0.31.0
anyio==4.4.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
boto3==1.34.142
botocore==1.34.142
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.0.1
clang==5.0
click==8.1.7
cohere==5.5.8
dataclasses-json==0.6.7
decoupler==1.6.0
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
Django==4.2.13
django-environ==0.11.2
docstring_parser==0.16
duckduckgo_search==6.1.12
exceptiongroup==1.2.1
fastavro==1.9.5
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.3.3
fsspec==2024.6.1
gast==0.4.0
google-api-core==2.19.1
google-auth==2.32.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.59.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
h2==4.1.0
h5py==3.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.0.0
jiter==0.5.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
keras==2.10.0
Keras-Preprocessing==1.1.2
langchain==0.2.7
langchain-anthropic==0.1.20
langchain-community==0.2.7
langchain-core==0.2.18
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langsmith==0.1.83
libclang==18.1.1
llvmlite==0.41.1
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.4.0
natsort==8.4.0
numba==0.58.1
numexpr==2.8.6
numpy==1.24.4
oauthlib==3.2.2
ollama==0.2.1
openai==1.35.13
opt-einsum==3.3.0
orjson==3.10.6
packaging==24.1
pandas==2.0.3
parameterized==0.9.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyreqwest_impersonate==0.4.9
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==0.21.1
python-telegram-bot==20.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
scipy==1.10.1
sentencepiece==0.2.0
shapely==2.0.5
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlparse==0.5.0
tenacity==8.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.42.3
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.18
virtualenv==20.26.3
Werkzeug==3.0.3
wikipedia==1.4.0
wrapt==1.12.1
yarl==1.8.2
zipp==3.19.2
